Refactor: 쿠키 기록 컴포넌트에 그룹명 삭제
