+++ src/pages/CookieRecordPage/LogSection/CookieLogList/CookieLogText/index.tsx
@@ -1,19 +1,21 @@
-import { Flex, Tag, Text } from '@chakra-ui/react'
+import { Flex, Image, Text } from '@chakra-ui/react'
+
+import Cookie1 from '@/assets/cookie1.svg'
 
 import { HintIcon } from './HintIcon'
 
 interface CookieLogTextProps {
-  groupName?: 'ALL' | string
   logContent: string
   hintCount: 0 | 1 | 2 | 3
   onClick: () => void
+  imageSrc?: string
 }
 
 export const CookieLogText = ({
-  groupName = 'ALL',
   logContent,
   hintCount,
   onClick,
+  imageSrc = Cookie1,
 }: CookieLogTextProps) => {
   return (
     <Flex
@@ -22,18 +24,8 @@ export const CookieLogText = ({
       onClick={onClick}
       _hover={{ cursor: 'pointer' }}
     >
-      <Tag
-        fontSize="x-small"
-        minWidth="fit-content"
-        height="fit-content"
-        borderRadius="full"
-        paddingY="0.3rem"
-        paddingX="0.6rem"
-        backgroundColor="brown.200"
-      >
-        {groupName}
-      </Tag>
-      <Flex gap={1} fontSize="small">
+      <Image src={imageSrc} width="20px" />
+      <Flex gap={1} fontSize="15px">
         <Text as="b">{logContent}</Text>
         <Text display="inline">으로 쿠키를 받았습니다.</Text>
       </Flex>
